check how it scales for special models
