+++ src/transformers/models/llama/modeling_llama.py
@@ -1311,7 +1311,9 @@ class LlamaForSequenceClassification(LlamaPreTrainedModel):
 
         loss = None
         if labels is not None:
-            loss = self.loss_function(logits=logits, labels=labels, pooled_logits=pooled_logits, config=self.config.todict())
+            loss = self.loss_function(
+                logits=logits, labels=labels, pooled_logits=pooled_logits, config=self.config.todict()
+            )
 
         if not return_dict:
             output = (pooled_logits,) + transformer_outputs[1:]
